--- src/components/AppLayout/FilterBar/index.ts
@@ -1,1 +1,0 @@
-export * from "./FilterBar";
